rename package to meet lint go standards
- terraform_docs_generator to terraformdocsgenerator
- openapi_terraform_docs_generator to openapiterraformdocsgenerator
