+++ utils/terraformdocsgenerator/openapiterraformdocsgenerator/openapi_terraform_provider_doc_generator_test.go
@@ -1,4 +1,4 @@
-package openapi_terraform_docs_generator
+package openapiterraformdocsgenerator
 
 import (
 	"errors"
