+++ utils/terraformdocsgenerator/openapiterraformdocsgenerator/openapi_terraform_provider_documentation.go
@@ -1,4 +1,4 @@
-package openapi_terraform_docs_generator
+package openapiterraformdocsgenerator
 
 import (
 	"io"
@@ -14,33 +14,34 @@ type TerraformProviderDocumentation struct {
 	ShowSpecialTermsDefinitions bool
 }
 
+// RenderHTML writes to the given writer argument the Terraform provider documentation
 func (t TerraformProviderDocumentation) RenderHTML(w io.Writer) error {
 	return t.renderZendeskHTML(w, TableOfContentsTmpl, ProviderInstallationTmpl, ProviderConfigurationTmpl, ProviderResourcesTmpl, DataSourcesTmpl, SpecialTermsTmpl)
 }
 
 // RenderZendeskHTML renders the documentation in HTML
 func (t TerraformProviderDocumentation) renderZendeskHTML(w io.Writer, tableOfContentsTemplate, providerInstallationTemplate, providerConfigurationTemplate, providerResourcesConfiguration, providerDatSourcesTemplate, specialTermsDefinitionsTemplate string) error {
-	err := Render(w, "TerraformProviderDocTableOfContents", tableOfContentsTemplate, t)
+	err := render(w, "TerraformProviderDocTableOfContents", tableOfContentsTemplate, t)
 	if err != nil {
 		return err
 	}
-	err = Render(w, "ProviderInstallation", providerInstallationTemplate, t.ProviderInstallation)
+	err = render(w, "ProviderInstallation", providerInstallationTemplate, t.ProviderInstallation)
 	if err != nil {
 		return err
 	}
-	err = Render(w, "ProviderConfiguration", providerConfigurationTemplate, t.ProviderConfiguration)
+	err = render(w, "ProviderConfiguration", providerConfigurationTemplate, t.ProviderConfiguration)
 	if err != nil {
 		return err
 	}
-	err = Render(w, "ProviderResources", providerResourcesConfiguration, t.ProviderResources)
+	err = render(w, "ProviderResources", providerResourcesConfiguration, t.ProviderResources)
 	if err != nil {
 		return err
 	}
-	err = Render(w, "DataSources", providerDatSourcesTemplate, t.DataSources)
+	err = render(w, "DataSources", providerDatSourcesTemplate, t.DataSources)
 	if err != nil {
 		return err
 	}
-	err = Render(w, "TerraformProviderDocSpecialTermsDefinitions", specialTermsDefinitionsTemplate, t)
+	err = render(w, "TerraformProviderDocSpecialTermsDefinitions", specialTermsDefinitionsTemplate, t)
 	if err != nil {
 		return err
 	}
