+++ utils/terraformdocsgenerator/openapiterraformdocsgenerator/openapi_terraform_provider_documentation_resources_test.go
@@ -1,4 +1,4 @@
-package openapi_terraform_docs_generator
+package openapiterraformdocsgenerator
 
 import (
 	"github.com/stretchr/testify/assert"
