+++ utils/terraformdocsgenerator/openapiterraformdocsgenerator/openapi_terraform_provider_documentation_test.go
@@ -1,4 +1,4 @@
-package openapi_terraform_docs_generator
+package openapiterraformdocsgenerator
 
 import (
 	"bytes"
@@ -114,32 +114,32 @@ func TestTerraformProviderDocumentation_RenderZendeskHTML_Errors(t *testing.T) {
 		{
 			name:                    "provider installation template error",
 			providerInstallTemplate: `{{.nonExistentVariable}}`,
-			expectedErr:             errors.New("template: ProviderInstallation:1:2: executing \"ProviderInstallation\" at <.nonExistentVariable>: can't evaluate field nonExistentVariable in type openapi_terraform_docs_generator.ProviderInstallation"),
+			expectedErr:             errors.New("template: ProviderInstallation:1:2: executing \"ProviderInstallation\" at <.nonExistentVariable>: can't evaluate field nonExistentVariable in type openapiterraformdocsgenerator.ProviderInstallation"),
 		},
 		{
 			name:                   "provider configuration template error",
 			providerConfigTemplate: `{{.nonExistentVariable}}`,
-			expectedErr:            errors.New("template: ProviderConfiguration:1:2: executing \"ProviderConfiguration\" at <.nonExistentVariable>: can't evaluate field nonExistentVariable in type openapi_terraform_docs_generator.ProviderConfiguration"),
+			expectedErr:            errors.New("template: ProviderConfiguration:1:2: executing \"ProviderConfiguration\" at <.nonExistentVariable>: can't evaluate field nonExistentVariable in type openapiterraformdocsgenerator.ProviderConfiguration"),
 		},
 		{
 			name:                      "provider resources template error",
 			providerResourcesTemplate: `{{.nonExistentVariable}}`,
-			expectedErr:               errors.New("template: ProviderResources:1:2: executing \"ProviderResources\" at <.nonExistentVariable>: can't evaluate field nonExistentVariable in type openapi_terraform_docs_generator.ProviderResources"),
+			expectedErr:               errors.New("template: ProviderResources:1:2: executing \"ProviderResources\" at <.nonExistentVariable>: can't evaluate field nonExistentVariable in type openapiterraformdocsgenerator.ProviderResources"),
 		},
 		{
 			name:                        "data sources template error",
 			providerDataSourcesTemplate: `{{.nonExistentVariable}}`,
-			expectedErr:                 errors.New("template: DataSources:1:2: executing \"DataSources\" at <.nonExistentVariable>: can't evaluate field nonExistentVariable in type openapi_terraform_docs_generator.DataSources"),
+			expectedErr:                 errors.New("template: DataSources:1:2: executing \"DataSources\" at <.nonExistentVariable>: can't evaluate field nonExistentVariable in type openapiterraformdocsgenerator.DataSources"),
 		},
 		{
 			name:                    "table of contents template error",
 			tableOfContentsTemplate: `{{.nonExistentVariable}}`,
-			expectedErr:             errors.New("template: TerraformProviderDocTableOfContents:1:2: executing \"TerraformProviderDocTableOfContents\" at <.nonExistentVariable>: can't evaluate field nonExistentVariable in type openapi_terraform_docs_generator.TerraformProviderDocumentation"),
+			expectedErr:             errors.New("template: TerraformProviderDocTableOfContents:1:2: executing \"TerraformProviderDocTableOfContents\" at <.nonExistentVariable>: can't evaluate field nonExistentVariable in type openapiterraformdocsgenerator.TerraformProviderDocumentation"),
 		},
 		{
 			name:                            "special terms definitions template error",
 			specialTermsDefinitionsTemplate: `{{.nonExistentVariable}}`,
-			expectedErr:                     errors.New("template: TerraformProviderDocSpecialTermsDefinitions:1:2: executing \"TerraformProviderDocSpecialTermsDefinitions\" at <.nonExistentVariable>: can't evaluate field nonExistentVariable in type openapi_terraform_docs_generator.TerraformProviderDocumentation"),
+			expectedErr:                     errors.New("template: TerraformProviderDocSpecialTermsDefinitions:1:2: executing \"TerraformProviderDocSpecialTermsDefinitions\" at <.nonExistentVariable>: can't evaluate field nonExistentVariable in type openapiterraformdocsgenerator.TerraformProviderDocumentation"),
 		},
 	}
